Experimenting with a flashlight light source
This is one potential method of concealing rooms and the enemies/items inside, so that players don't know where everything is.

The walls had to be made taller so that the light doesn't go over the walls and into the rooms, and enemies had to be made shorter so that light could hit the top of the enemy object to show their color. Otherwise the enemies would look like black shadowy circles.