Add Command Terminal to End Game
Created the command terminal game object that allows the player to capture the ship and end the game. Once the player has gotten the necessary key for the command center, they can enter the room, go to the terminal, and press space to capture the ship and load the game end screen. It also shows a popup message telling the player to press space.

The terminal loads the _Game_Win scene, which should be modified with the actual screen visuals, as it is currently just an empty scene.