Add Player Lives and Game Over Screen
The player now has three lives and if the player dies with lives remaining, then the scene will wait for a few seconds then load the next scene so the player can try again.

The remaining player lives are displayed in the UI.

If the player dies with no lives remaining, then they see a game over screen with a play again button.